renamed the *TestCase classes to *Demo
Solves Issue: #3730

---

I hereby agree to the terms of the JUnit Contributor License Agreement.
